XPCOM API difference between Firefox and SeaMonkey in utils.js. Added conditional to resolve for SeaMonkey.
JavaScript console lockup due in part to lack of Extensions manager present in SeaMonkey.  This is only a partial fix. Philip Chee will need to enlighten us on why it still locks when Greasemonkey is installed in a local profile instead of the global profile.

Modified build.sh to account for static version requirement for SeaMonkey since Extension Manager is not present.